+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 ### 4.10.9 (3/3/2017 - )
 * lib: update kafka to 0.10.2.0, es to 5.2.2
+* kafka: simplify listener, warn if took too long to consume message
 
 ### 4.10.8 (2/28/2017 - 3/2/2017)
 * api: validate webservice impl method should not have @PathParam()